Update client\package.json (See description for more information)
I updated the versions in the client\package.json so that the React app would run properly. For example, in index.js, the code with have on line 2 (import ReactDOM from 'react-dom/client') is only compatible with React version 18 and up, so I updated React from 17.0.1 to 18.2.0. I also removed the apollo/client, graphql, and jwt-decode dependencies because these are only required in the server json file. I also removed the testing dependencies because we are likely not going to use them. Finally, I removed the line about the proxy because it was giving me errors, and I don't think we need this line of code.